latest version

+++ src/get_max_flow_paths.c
@@ -46,7 +46,7 @@ t_vertex *sink, t_vertex *sink_adj)
 }
 
 static t_array	*save_flow_paths(t_graph *graph, t_vertex *src,
-t_vertex *sink, int count)
+t_vertex *sink, size_t count)
 {
 	t_array	*paths;
 	t_array	*path;
@@ -84,6 +84,6 @@ t_array	*get_max_flow_paths(t_graph *graph)
 	max_flow = max_flow_edmonds_karp(graph, graph->source, graph->sink);
 	if (max_flow <= 0)
 		return (NULL);
-	paths = save_flow_paths(graph, graph->source, graph->sink, max_flow);
+	paths = save_flow_paths(graph, graph->source, graph->sink, (size_t)max_flow);
 	return (paths);
 }
